% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfields-package.r
\docType{package}
\name{ggfields-package}
\alias{ggfields}
\alias{ggfields-package}
\title{ggfields: Add Vector Field Layers to Ggplots}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Add vector field layers to ggplots. Ideal for visualising wind speeds, water currents, electric/magnetic fields, etc. Accepts data.frames simple features (sf) and spatiotemporal arrays (stars) objects as input. Vector fields are depicted as arrows starting at specified locations, and with specified angles and radii.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pepijn-devries.github.io/ggfields/}
  \item \url{https://github.com/pepijn-devries/ggfields/}
}

}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

}
\keyword{internal}
