% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{seawatervelocity}
\alias{seawatervelocity}
\title{A small subset of the global ocean physics analysis and forecast product}
\format{
A \link[stars:st_as_stars]{stars} object with \code{x}, \code{y}, \code{depth} and \code{time} dimensions. It has the
attributes \code{vo} (northward seawater velocity [m/s]) and \code{uo} (eastward seawater velocity [m/s]).
}
\description{
A small subset of ocean currents data retrieved with \link[CopernicusMarine:CopernicusMarine-package]{CopernicusMarine}
from the source listed below serving as an example.
}
\examples{
data("seawatervelocity")
}
\references{
E.U. Copernicus Marine Service Information; Global Ocean Physics Analysis and Forecast - GLOBAL_ANALYSISFORECAST_PHY_001_024 (2016-10-14). \doi{DOI:10.48670/moi-00016}
}
