% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fields.r
\docType{data}
\name{StatFields}
\alias{StatFields}
\alias{stat_fields}
\title{Stat method for geom_fields}
\format{
An object of class \code{StatFields} (inherits from \code{StatSf}, \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatFields

stat_fields(
  mapping = NULL,
  data = NULL,
  geom = "fields",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{Can be one of four things:
\itemize{
\item \code{NULL}: in that case data from the parent \link[ggplot2:ggplot]{ggplot} call is
inherited.
\item \code{data.frame}: you need to assign the \code{x} and \code{y} aesthetics.
\item \link[sf:st_as_sf]{sf object}: it should contain a \code{geometry} column with only \code{POINT} geometries.
\item \link[stars:st_as_stars]{stars object}: it will be converted automatically to an \code{sf} object.
}}

\item{geom}{The layer type for which the data is prepared. In this case \code{"fields"}.}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
Returns a \link[ggplot2:layer]{layer} that can be further modified by \code{\link[=geom_fields]{geom_fields()}}.
}
\description{
Prepares data before being handled by \code{\link[=geom_fields]{geom_fields()}}
}
\examples{
stat_fields()
}
\author{
Pepijn de Vries
}
\keyword{datasets}
